\name{extract-methods}
\docType{methods}
\alias{extract}
\alias{extract-methods}
\alias{extract,ANY-method}
\title{Generic function \code{extract}}
\description{
 Provides generic function \code{extract} currently used in 
packages \code{aws} and \code{dti}.
}
\section{Methods}{
\describe{Method \code{extract} is used to extract information from objects 
defined in packages \code{aws} and \code{dti}. Here only the generic function is 
defined. The method usually takes an argument \code{what='vector of characters'}
defining which information is needed and returns a list with component names corresponding
to the entries in \code{what}.

\item{\code{signature(x = "ANY")}}{
%%  ~~describe this method here~~
}
}}
\keyword{methods}
