\name{risk-methods}
\docType{methods}
\alias{risk}
\alias{risk-methods}
\alias{risk,ANY-method}
\title{Generic function \code{risk}}
\description{
Provides generic function \code{risk} currently used in 
package \code{aws}.
}
\section{Methods}{
\describe{Defines a method used in package \code{aws} to evaluate the 
quality of smoothing results. See help in package \code{aws} for details.

\item{\code{signature(y = "ANY")}}{
%%  ~~describe this method here~~
}
}}
\keyword{methods}
